bicep/requirements.txt
azure-cli==2.54.0
python-dotenv==1.0.0
azure-core==1.29.4
